Also move the fermion device to the DjangoDevice
