+++ pennylane_ls/__init__.py
@@ -4,6 +4,6 @@ from .SingleQuditOps import load, rLx, Lz, rLz, rLz2
 from .MultiQuditDevice import MultiQuditDevice
 
 from .FermionOps import Load, Hop, Inter, Phase, ParticleNumber, PauliZ, HartreeFock
-from .FermionDevice import FermionDevice
+from .fermion_device import FermionDevice
 
 from ._version import __version__
