+++ pennylane_ls/django_device.py
@@ -3,6 +3,9 @@ Define the base class for communication with the Django API as laid out for
 `labscript-qc`
 """
 
+import time
+import json
+import requests
 from pennylane import Device
 
 
@@ -24,10 +27,54 @@ class DjangoDevice(Device):
         """
         The initial part.
         """
-        super().__init__(wires=1, shots=shots)
+        super().__init__(wires=wires, shots=shots)
         self.username = username
         self.password = password
         self.blocking = blocking
         self.job_id = job_id
         self.url_prefix = url
         self.job_payload = {}
+
+    def check_job_status(self) -> str:
+        """
+        Check remotely if the job was done already.
+        """
+        status_payload = {"job_id": self.job_id}
+        url = self.url_prefix + "get_job_status/"
+        status_response = requests.get(
+            url,
+            params={
+                "json": json.dumps(status_payload),
+                "username": self.username,
+                "password": self.password,
+            },
+        )
+        job_status = (status_response.json())["status"]
+        return job_status
+
+    def wait_till_done(self):
+        """
+        The waiting function that blocks the program
+        """
+        while True:
+            time.sleep(2)
+            job_status = self.check_job_status()
+            if job_status == "DONE":
+                break
+
+    def pre_apply(self):
+        '''
+        Set up the necessary dictonaries that will be later send to the server.
+        '''
+        self.reset()
+        self.job_payload = {
+            "experiment_0": {"instructions": [], "num_wires": 1, "shots": self.shots},
+        }
+
+    @property
+    def operations(self):
+        return set(self._operation_map.keys())
+
+    @property
+    def observables(self):
+        return set(self._observable_map.keys())
